# server/openapi.yaml
openapi: 3.0.3
info:
  title: AIBillingSystem API
  description: |
    API for the AIBillingSystem - a comprehensive medical billing and practice management system
    designed for therapy practices. Features include patient management, claims processing,
    appointment scheduling, analytics, and AI-powered billing optimization.

    ## Authentication
    Most endpoints require authentication via session-based auth (Replit Auth).
    Protected endpoints will return 401 Unauthorized if not authenticated.

    ## Multi-Tenancy
    The system supports multiple practices. Users are assigned to a practice and can only
    access data within their authorized practice (except admin users who have cross-practice access).

    ## HIPAA Compliance
    This API handles Protected Health Information (PHI). All access is logged for audit purposes.
  version: 1.0.0
  contact:
    name: AIBillingSystem Support
  license:
    name: Proprietary
servers:
  - url: /api
    description: API base path
...
paths:
...
  /claims:
    get:
      tags:
        - Claims
      summary: List all claims
      description: Returns all claims for the authorized practice
      operationId: listClaims
      security:
        - sessionAuth: []
      responses:
        '200':
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Claim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Claims
      summary: Create a new claim
      description: Creates a new claim with optional AI optimization based on session SOAP notes
      operationId: createClaim
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreate'
      responses:
        '200':
          description: Claim created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Claim created successfully
                  claim:
                    $ref: '#/components/schemas/Claim'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication via Replit Auth
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Unauthorized
...
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Invalid request parameters
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: An unexpected error occurred
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
...
    Claim:
      type: object
      properties:
        id:
          type: integer
        practiceId:
          type: integer
        patientId:
          type: integer
        sessionId:
          type: integer
        claimNumber:
          type: string
        insuranceId:
          type: integer
        totalAmount:
          type: string
          format: decimal
        submittedAmount:
          type: string
          format: decimal
        paidAmount:
          type: string
          format: decimal
        expectedAmount:
          type: string
          format: decimal
        status:
          type: string
          enum: [draft, submitted, paid, denied, appeal, optimized]
        submittedAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        denialReason:
          type: string
        aiReviewScore:
          type: string
          format: decimal
        aiReviewNotes:
          type: string
        clearinghouseClaimId:
          type: string
        clearinghouseStatus:
          type: string
          enum: [accepted, rejected, pending]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ClaimCreate:
      type: object
      required:
        - patientId
        - totalAmount
      properties:
        patientId:
          type: integer
        insuranceId:
          type: integer
        sessionId:
          type: integer
        totalAmount:
          type: number
          format: decimal
        submittedAmount:
          type: number
          format: decimal
...
security:
  - sessionAuth: []